Fix lazy tolist cache metadata
