+++ torch/_dynamo/variables/torch.py
@@ -1770,13 +1770,9 @@ class TorchInGraphFunctionVariable(BaseTorchVariable):
             *args: VariableTracker,
             **kwargs: VariableTracker,
         ) -> VariableTracker | None:
-            from .tensor import materialize_tensor_tolist_arg
+            from .tensor import materialize_tensor_tolist_args_kwargs
 
-            args = tuple(materialize_tensor_tolist_arg(arg, tx) for arg in args)
-            kwargs = {
-                key: materialize_tensor_tolist_arg(arg, tx)
-                for key, arg in kwargs.items()
-            }
+            args, kwargs = materialize_tensor_tolist_args_kwargs(tx, args, kwargs)
 
             def check_any_unspec(x: VariableTracker) -> bool:
                 # NB: This includes UnspecializedPythonVariable
@@ -2635,7 +2631,7 @@ class TorchInGraphFunctionVariable(BaseTorchVariable):
     ) -> "VariableTracker":
         from . import SymNodeVariable
         from .builder import wrap_fx_proxy
-        from .tensor import materialize_tensor_tolist_arg
+        from .tensor import materialize_tensor_tolist_args_kwargs
 
         if self.kind == AllowInGraphKind.NONSTRICT_TRACE:
             return self._call_nonstrict_traceable_function(tx, args, kwargs)
@@ -2810,10 +2806,7 @@ For now, dynamo will explicitly graph break when it encounters user code with th
             if tx.fake_mode and tx.fake_mode.shape_env:
                 ctx = tx.fake_mode.shape_env.ignore_fresh_unbacked_symbols
 
-        args = [materialize_tensor_tolist_arg(arg, tx) for arg in args]
-        kwargs = {
-            key: materialize_tensor_tolist_arg(arg, tx) for key, arg in kwargs.items()
-        }
+        args, kwargs = materialize_tensor_tolist_args_kwargs(tx, args, kwargs)
 
         with ctx():
             tensor_variable = wrap_fx_proxy(
